+++ CHANGELOG.md
@@ -28,6 +28,7 @@
     * instance creation takes a string value
     * all methods `to*()` are additionally available as non-static `as*()` methods except `toType()`
   * Added `stubbles\lang\Properties::parse()` which returns an instance of `stubbles\lang\Parse`
+  * Added `stubbles\lang\reflect\annotation\Annotation::parse()` which returns an instance of `stubbles\lang\Parse`
 
 
 4.1.4 (2014-08-11)